feat(strawberry-poc): add BigInt scalar for file_size >2GB
Production file_size values can exceed GraphQL's 32-bit Int limit
(legacy schema used a BigInt scalar). Add a Strawberry BigInt scalar
in models/common.py and switch every file_size declaration to use it:
the two interfaces (FileInterface, InversionSolutionInterface) and
all 9 CreateXInput types across the file/solution model files.

Removes the xfail marker on test_create_file_with_large_file_size_bigint;
it now passes with file_size: 5_000_000_000. 151 tests pass (was 150+1xfail).

+++ spike/strawberry_poc/models/aggregate_inversion_solution.py
@@ -11,7 +11,7 @@ from strawberry.types import Info
 from data.dynamo import create_file, get_file, list_files
 from data.models import AggregateInversionSolutionData
 
-from .common import AggregationFn, KeyValuePair, KeyValuePairInput, _try_enum
+from .common import AggregationFn, BigInt, KeyValuePair, KeyValuePairInput, _try_enum
 from .file_interface import FileInterface
 from .inversion_solution import LabelledTableRelation, _ltr_from_dict
 from .inversion_solution_interface import InversionSolutionInterface
@@ -100,7 +100,7 @@ class CreateAggregateInversionSolutionInput:
     aggregation_fn: AggregationFn
     produced_by: strawberry.ID | None = None
     md5_digest: str | None = None
-    file_size: int | None = None
+    file_size: BigInt | None = None
     meta: list[KeyValuePairInput] | None = None
     created: str | None = None
     predecessors: list[PredecessorInput] | None = None
