+++ spike/strawberry_poc/models/inversion_solution_nrml.py
@@ -11,7 +11,7 @@ from strawberry.types import Info
 from data.dynamo import create_file, get_file, list_files
 from data.models import InversionSolutionNrmlData
 
-from .common import KeyValuePair, KeyValuePairInput
+from .common import BigInt, KeyValuePair, KeyValuePairInput
 from .file_interface import FileInterface
 from .predecessor import PredecessorInput
 from .predecessors_interface import PredecessorsInterface
@@ -61,7 +61,7 @@ class CreateInversionSolutionNrmlInput:
     file_name: str
     source_solution: strawberry.ID | None = None
     md5_digest: str | None = None
-    file_size: int | None = None
+    file_size: BigInt | None = None
     meta: list[KeyValuePairInput] | None = None
     created: str | None = None
     predecessors: list[PredecessorInput] | None = None
